@@ -47,25 +47,27 @@
     "!*/__tests__"
   ],
   "dependencies": {
-    "@loopback/authentication": "^1.1.1",
-    "@loopback/boot": "^1.1.2",
-    "@loopback/context": "^1.8.1",
-    "@loopback/core": "^1.2.1",
-    "@loopback/openapi-v3": "^1.3.2",
-    "@loopback/repository": "^1.2.1",
-    "@loopback/rest": "^1.9.1",
-    "@loopback/rest-explorer": "^1.1.13",
-    "@loopback/service-proxy": "^1.1.1",
+    "@loopback/authentication": "^2.0.2",
+    "@loopback/boot": "^1.2.6",
+    "@loopback/context": "^1.14.0",
+    "@loopback/core": "^1.6.3",
+    "@loopback/openapi-v3": "^1.3.10",
+    "@loopback/repository": "^1.5.4",
+    "@loopback/rest": "^1.11.1",
+    "@loopback/rest-explorer": "^1.1.21",
+    "@loopback/service-proxy": "^1.1.9",
+    "jsonwebtoken": "^8.5.1",
+    "passport": "^0.4.0",
     "passport-jwt": "^4.0.0"
   },
   "devDependencies": {
-    "@loopback/build": "^1.4.0",
-    "@loopback/testlab": "^1.2.1",
+    "@loopback/build": "^1.5.3",
+    "@loopback/testlab": "^1.2.8",
     "@loopback/tslint-config": "^2.0.3",
-    "@types/node": "^10.11.2",
+    "@types/node": "^10.14.6",
     "@types/passport": "^1.0.0",
     "@types/passport-jwt": "^3.0.1",
-    "tslint": "^5.12.0",
-    "typescript": "^3.3.1"
+    "tslint": "^5.16.0",
+    "typescript": "^3.4.5"
   }
 }
