@@ -3,13 +3,8 @@ import {UserRole} from '../models';
 import {MemDataSource} from '../datasources';
 import {inject} from '@loopback/core';
 
-export class UserRoleRepository extends DefaultCrudRepository<
-  UserRole,
-  typeof UserRole.prototype.id
-> {
-  constructor(
-    @inject('datasources.mem') dataSource: MemDataSource,
-  ) {
+export class UserRoleRepository extends DefaultCrudRepository<UserRole, typeof UserRole.prototype.id> {
+  constructor(@inject('datasources.mem') dataSource: MemDataSource) {
     super(UserRole, dataSource);
   }
 }
